app/src/test/java/com/bignerdranch/android/beatbox/SoundViewModelTest.java
package com.bignerdranch.android.beatbox;

import org.junit.Before;
import org.junit.Test;

import static org.hamcrest.Matchers.is;
import static org.junit.Assert.*;
import static org.mockito.Mockito.mock;
import static org.mockito.Mockito.verify;

/**
 * Created by Mateusz on 2017-04-02.
 */
public class SoundViewModelTest {

    private BeatBox mBeatBox;
    private Sound mSound;
    private SoundViewModel mSubject;

    @Before
    public void setUp() throws Exception {

        mBeatBox = mock(BeatBox.class);
        mSound = new Sound("assetPath");
        mSubject = new SoundViewModel(mBeatBox);
        mSubject.setSound(mSound);
    }

    @Test
    public void exposesSoundNameAsTitle() {
        assertThat(mSound.getmName(), is(mSubject.getTitle()));
    }

    @Test public void callsBeatBoxPlayOnButtonClicked() {

        mSubject.onButtonClicked();
        verify(mBeatBox).play(mSound);


    }

}